# task4/repo-stat/api/internal/controller/http/v1/dto.go
package v1

import "time"

// RepositoryResponse represents a successful response
type RepositoryResponse struct {
	FullName        string    `json:"full_name" example:"go"`
	Description     string    `json:"description" example:"The Go programming language"`
	StargazersCount int       `json:"stars" example:"123456"`
	ForksCount      int       `json:"forks" example:"12345"`
	CreatedAt       time.Time `json:"created_at" example:"2014-08-19T22:33:41Z"`
	HTMLURL         string    `json:"html_url" example:"https://github.com/golang/go"`
}

type SubscriptionResponse struct {
	Owner string `json:"owner" example:"google"`
	Repo  string `json:"repo" example:"go"`
}

// HealthResponse represents a services status response
type HealthResponse struct {
	Status   string          `json:"status" example:"ok"`
	Services []ServiceStatus `json:"services"`
}

// ServiceStatus represents a service status
type ServiceStatus struct {
	Name   string `json:"name" example:"processor"`
	Status string `json:"status" example:"up"`
}

// ErrorResponse represents a error response
type ErrorResponse struct {
	Error   string `json:"error" example:"Not Found"`
	Message string `json:"message" example:"Repository not found"`
}